+++ src/pwa/stores/connection/custom.js
@@ -1,6 +1,6 @@
 import { values, observable } from 'mobx';
 import { types, getParent } from 'mobx-state-tree';
-import { flatten } from 'lodash';
+import { flatten } from 'lodash-es';
 import { Total, Page } from './list';
 import { pageShape } from './list-shape';
 import { extract } from './utils';
@@ -8,11 +8,11 @@ import { extract } from './utils';
 const Custom = types
   .model('Custom')
   .props({
-    name: types.identifier(types.string),
+    name: types.identifier,
     pageMap: types.optional(types.map(Page), {}),
     total: types.optional(Total, {}),
     url: types.optional(types.string, '/'),
-    params: types.optional(types.frozen, {}),
+    params: types.frozen(),
   })
   .views(self => ({
     get isReady() {