+++ src/pwa/stores/index.js
@@ -1,5 +1,5 @@
 /* eslint-disable func-names */
-import { flow, types } from 'mobx-state-tree';
+import { types } from 'mobx-state-tree';
 import * as connection from './connection';
 import * as router from './router';
 import * as history from './router/history';
@@ -12,14 +12,6 @@ const Connection = types
   .props(router.props)
   .views(router.views)
   .actions(router.actions)
-  .actions(history.actions)
-  .actions(self => ({
-    beforeSsr: flow(function*() {
-      yield self.fetchHeadContent();
-    }),
-    afterCsr() {
-      self.replaceFirstUrl();
-    },
-  }));
+  .actions(history.actions);
 
 export default Connection;